langchain/requirements.txt
torch
sentence-transformers
faiss-cpu
langchain
openai
deeplake
python-dotenv